Corrections to Interrupts - i16.h
Dropped max Interrupt value to uint8_max as I realised that handling uint16_max with 20 bytes per vector (with assumption that instructions are at least 4 bytes, meaning 5 instructions per vector) would consume ~1MiB where as 256 would consume ~5KiB.

+++ emu/i16.h
@@ -91,7 +91,7 @@ public:
         registers[FLAG] == 0x8000;
     }
 
-    std::array<bool, UINT16_MAX + 1> interrupts;
+    std::array<bool, UINT8_MAX + 1> interrupts;
     bool interruptActive;
 
     std::array<uint16_t, UINT8_MAX + 1> ports;
